+++ PropertyManagerFL.Core/Entities/Concelho.cs
@@ -5,5 +5,5 @@ public class Concelho
     public int CodConcelho { get; set; }
     public string? Descricao { get; set; }
     public int IdDistrito { get; set; }
-    public float Coeficiente { get; set; }
+    public decimal Coeficiente { get; set; }
 }
